# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Product Validation Platform API",
    "version": "1.0.0",
    "description": "Complete API documentation for the AI-powered product research and validation platform"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/auth/signup": {
      "post": {
        "summary": "Register new user",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "minLength": 8 }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "object",
                      "properties": {
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": { "type": "string" },
                            "email": { "type": "string" }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation error" },
          "409": { "description": "Email already registered" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User login",
        "tags": ["Authentication"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string" },
                  "code": { "type": "string", "description": "2FA code if enabled" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful" },
          "401": { "description": "Invalid credentials or 2FA required" }
        }
      }
    },
    "/validations": {
      "get": {
        "summary": "List user validations",
        "tags": ["Validations"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of validations",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Validation" }
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" }
        }
      },
      "post": {
        "summary": "Create new validation",
        "tags": ["Validations"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["ideaDescription"],
                "properties": {
                  "ideaDescription": { "type": "string", "minLength": 10 },
                  "targetAudience": { "type": "string" },
                  "industry": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Validation created" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/validations/{id}": {
      "get": {
        "summary": "Get validation details",
        "tags": ["Validations"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Validation details" },
          "404": { "description": "Validation not found" }
        }
      },
      "delete": {
        "summary": "Delete validation",
        "tags": ["Validations"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": { "description": "Deleted successfully" },
          "404": { "description": "Validation not found" }
        }
      }
    },
    "/validations/{id}/competitors": {
      "post": {
        "summary": "Add competitor analysis",
        "tags": ["Competitors"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Competitor" }
            }
          }
        },
        "responses": {
          "201": { "description": "Competitor added" }
        }
      }
    },
    "/billing": {
      "get": {
        "summary": "Get billing info",
        "tags": ["Billing"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Billing information" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Validation": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userId": { "type": "string" },
          "ideaDescription": { "type": "string" },
          "targetAudience": { "type": "string" },
          "status": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Competitor": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "url": { "type": "string" },
          "description": { "type": "string" },
          "features": { "type": "array", "items": { "type": "object" } },
          "pricing": { "type": "array", "items": { "type": "object" } },
          "strengths": { "type": "array", "items": { "type": "string" } },
          "weaknesses": { "type": "array", "items": { "type": "string" } },
          "swot": { "type": "object" }
        }
      }
    }
  }
}